Refactor Stripe webhook for Firebase Functions v2
